+++ mopaq/tests/security/mod.rs
@@ -4,3 +4,4 @@ mod crc;
 mod crc_generation;
 mod crypto;
 mod hash;
+mod signature;